Shift a lot of stuff onto the 9S12. 9S12 uses fixed point rather than floating point. Does Riemann sums with rectangles, not trapezoids. Does a running tally on displacement since the last packet.  Sends packets at rate of 10 Hz.
Packet contents are now half the size, discounting the 3 byte overhead. (dropped tick number, lateral movement).

+++ branches/INS/MCU/Command.h
@@ -100,7 +100,7 @@ void Command_processCmd(uint8_t* cmd, uint16_t size);
  * Sends a compass packet to the x86.
  */
 void Command_sendCompassPacket(uint8_t sensorType, uint16_t compassHeading);
-void Command_sendAccelerometerPacket(uint8_t tickNumber, uint16_t x, uint16_t y, uint16_t gyro);
+void Command_sendAccelerometerPacket(uint16_t tickNumber, int16_t x, int16_t y, int16_t gyro);
 
 void Command_startSendingData(void);
 void Command_stopSendingData(void);
